+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### [1.3.1] - 2026-08-23
 
-- Add List-Unsubscribe headers to DKIM config (#30)
+- config: sign List-Unsubscribe headers (RFC 8058) #30
+- config: sign Content-Type & Content-Transfer-Encoding
+- doc(README): note to diff dkim.ini when upgrading
 
 ### [1.3.0] - 2026-07-31